Update API documentation and assets
This commit includes several updates to the API documentation, including formatting improvements for method signatures and the addition of new security event retrieval methods. Additionally, the theme of the documentation has been changed from "linden" to "willow," and color values have been updated for better visual consistency. The favicon has been replaced with a new design, and the README file has been removed as it is no longer needed.